debug

+++ scripts/build_audio_cache.py
@@ -8,6 +8,7 @@ import sys
 import time
 from transformers import AutoTokenizer
 from collections import defaultdict
+from itertools import product
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
@@ -44,14 +45,16 @@ def save_bucket(samples, bucket, cache_dir, bucket_id):
     """
     Save a bucket of embeddings to disk and update sample metadata.
     """
-    pt_path = os.path.join(cache_dir, f"bucket_{bucket_id:06d}.pt")
+    pt_path = os.path.join(cache_dir, f"bucket_bs{len(bucket)}_{bucket_id:06d}.pt")
     tmp_path = pt_path + ".tmp"
 
     indices = [idx for idx, _ in bucket]
     embs = torch.stack([emb for _, emb in bucket])  # (B, T, D)
 
-    torch.save({"audio_embs": embs}, tmp_path, _use_new_zipfile_serialization=False)
-    os.replace(tmp_path, pt_path)
+    # do not save if file already exists (can happen if multiple processes are running this script with same cache_dir)
+    if not os.path.exists(pt_path):
+        torch.save({"audio_embs": embs}, tmp_path, _use_new_zipfile_serialization=False)
+        os.replace(tmp_path, pt_path)
 
     # update sample metadata
     for i, idx in enumerate(indices):
@@ -66,6 +69,12 @@ def save_sorted_samples(samples, embedder_path, batch_size, bucket_size, cache_d
     bucket_id = 0
     t_embedding = 0.0
     t_saving = 0.0
+
+    if os.path.exists(os.path.join(cache_dir, "meta.json")):
+        logger.info(f"Cache directory {cache_dir} already contains meta.json, skipping embedding and saving")
+        return
+
+    os.makedirs(cache_dir, exist_ok=True)
 
     torch_dtype = getattr(torch, dtype)
 
@@ -112,6 +121,19 @@ def save_sorted_samples(samples, embedder_path, batch_size, bucket_size, cache_d
     logger.info(f"Saved {len(samples)} embeddings in {bucket_id} buckets dir={cache_dir}")
     logger.info(f"Embedding time = {t_embedding:.2f}s, Saving time = {t_saving:.2f}s")
 
+    # Save meta.json
+    meta_path = os.path.join(args.cache_dir, "meta.json")
+    info = {
+        "json_path": args.json_path,
+        "cache_dir": args.cache_dir,
+        "embedder_path": args.embedder_path,
+        "tokenizer_path": args.tokenizer_path,
+        "dtype": args.dtype,
+        "bucket_size": args.bucket_size,
+    }
+    with open(meta_path, "w", encoding="utf-8") as f:
+        json.dump({"info": info, "samples": samples}, f, ensure_ascii=False)
+    logger.info(f"Saved {meta_path}")
 
 
 
@@ -147,6 +169,10 @@ if __name__ == "__main__":
     # Compute tokenized lengths
     key2sample = defaultdict(dict)
 
+    splits = set()
+    slangs = set()
+    tlangs = set()
+
     for s in tqdm(samples, total=len(samples), desc="Tokenizing text", unit=" sample"):
         audio_file = s.get("audio_file", "")
 
@@ -175,6 +201,10 @@ if __name__ == "__main__":
         if not isinstance(text, str) or not text.strip():
             continue
 
+        splits.add(split)
+        slangs.add(slang)
+        tlangs.add(tlang)
+
         ids = tokenizer(text, padding=False, truncation=False, add_special_tokens=False)["input_ids"]
         key2sample[audio_file] = {"audio_file": audio_file, "text": text, "ids": ids, "slang": slang, "tlang": tlang, "split": split, "len": len(ids)}
 
@@ -183,31 +213,20 @@ if __name__ == "__main__":
         sys.exit(0)
 
     logger.info(f"Found {len(samples)} unique audio files with valid transcriptions after tokenization.")
-
-    # Sort samples by tokenized length (shortest → longest)
-    samples = list(key2sample.values())
-    samples.sort(key=lambda x: x["len"])
-    logger.info(f"Sorted {len(samples)} samples by tokenized length. Shortest len={samples[0]['len']}, longest len={samples[-1]['len']}")
-
-    sys.exit()
-    
-    os.makedirs(args.cache_dir, exist_ok=True)
+    logger.info(f"Splits: {splits}")
+    logger.info(f"slangs: {slangs}")
+    logger.info(f"tlangs: {tlangs}")
 
     #################################################################################
     ### Save audio embeddings in bucketed .pt files #################################
     #################################################################################
-    save_sorted_samples(samples, args.embedder_path, args.batch_size, args.bucket_size, args.cache_dir, args.device, args.dtype)
 
-    # Save meta.json
-    meta_path = os.path.join(args.cache_dir, "meta.json")
-    info = {
-        "json_path": args.json_path,
-        "cache_dir": args.cache_dir,
-        "embedder_path": args.embedder_path,
-        "tokenizer_path": args.tokenizer_path,
-        "dtype": args.dtype,
-        "bucket_size": args.bucket_size,
-    }
-    with open(meta_path, "w", encoding="utf-8") as f:
-        json.dump({"info": info, "samples": samples}, f, ensure_ascii=False)
-    logger.info(f"Saved {meta_path}")
+    combinations = list(product(splits, slangs, tlangs))
+    logger.info(f"Split-Slang-Tlang combinations: {combinations}")
+    for split, slang, tlang in combinations:
+        combinations_samples = [s for s in key2sample.values() if (args.split is None or s['split'] == split) and (args.slang is None or s['slang'] == slang) and (args.tlang is None or s['tlang'] == tlang)]
+        combinations_samples.sort(key=lambda x: (x["len"], x["audio_file"])) # sort by tokenized length, then by audio file name for tie-breaking
+        logger.info(f"Combination (split={split}, slang={slang}, tlang={tlang}): {len(combinations_samples)} samples")
+        odir = os.path.join(args.cache_dir, f"{split}/{slang}/{tlang}")
+        save_sorted_samples(combinations_samples, args.embedder_path, args.batch_size, args.bucket_size, odir, args.device, args.dtype)
+
